mc.py: restructure
I splitted out most of the FSI code for mc into a new 'stashutils'

package. The dropbox-account-system has also been moved into
'stashutils'. 'mc' supports now more than one Dropbox user.
