+++ lib/stashutils/__init__.py
@@ -1,0 +1,1 @@
+"""utility functions and modules for StaSh-scripts"""
