+++ lib/stashutils/fsi/__init__.py
@@ -1,0 +1,5 @@
+"""
+This package contains classes used for accessing filesystems or servers.
+These classes have a common interface, which means they can easily be
+switched.
+"""
